COMPLETE: File reorganization and fixes
**Major File Reorganization:**
- Created functional directory structure:
  - app/cars/ - Car management (details, edit, list, factory, manage, identify)
  - app/contact/ - Communication features (contact forms, owner messaging)
  - app/reports/ - Analytics and statistics
  - app/cars/actions/ - Car-related AJAX endpoints

**Backward Compatibility:**
- Created 301 redirect files for all moved pages
- All existing URLs continue to work seamlessly

**JavaScript & Asset Fixes:**
- Fixed statistics page Google Charts data injection timing
- Fixed car details page map display and variable conflicts
- Updated all relative paths after file moves
- Fixed image paths in cars/identify.php for example images

**Security & Code Quality:**
- Added new directories to UserSpice path security management
- Fixed PHP 8.1+ deprecation warning in _edit_car_2.php
- Updated AJAX endpoints and form action URLs
- Maintained all existing functionality

All pages now work correctly in their new organized structure.
