+++ userimages/.htaccess
@@ -1,0 +1,4 @@
+<FilesMatch ".(jpg|gif|png)$">
+    Order Deny,Allow
+    Allow from all
+</FilesMatch>
